Remove duplicate commits
Handled the case where duplicate commit dictionaries were being created.
Because a user can have access to projects that are forks of an original
project, the same commit can be found in multiple projects. To handle
this, commit hashes are stored in a set and checked before adding a new
commit dictionary to the list of commits. To ensure the earliest version
of the commit is stored, projects are sorted in ascending order by
creation date, then processed for commits that were made by the user.
Forked projects are created after the original project, so in processing
the projects in ascending order by date, we ensure that the earliest
version of the commit is stored.